+++ packages/cache-url/lib/AmpCurlUrlGenerator.js
@@ -17,6 +17,11 @@
 import Url from 'url-parse';
 import punycode from 'punycode';
 
+// Our imports that are dynamically filtered
+// by rollup
+import ampCurlUrlBrowserSha256 from './AmpCurlUrlBrowserSha256';
+import ampCurlUrlNodeSha256 from './AmpCurlUrlNodeSha256';
+
 /** @type {string} */
 const LTR_CHARS =
   'A-Za-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02B8\u0300-\u0590\u0800-\u1FFF' +
@@ -138,42 +143,10 @@ function constructFallbackCurlsCacheDomain_(domain) {
  */
 function sha256_(str) {
   if (typeof window !== 'undefined') {
-    // Transform the string into an arraybuffer.
-    const buffer = new TextEncoder('utf-8').encode(str);
-    return crypto.subtle.digest('SHA-256', buffer).then((hash) => {
-      return hex_(hash);
-    });
+    return ampCurlUrlBrowserSha256(str);
   } else {
-    const buffer = Buffer.from(str, 'utf-8');
-    const crypto = require('crypto');
-    return new Promise((resolve) => {
-      const sha256 = crypto.createHash('sha256').update(buffer).digest('hex');
-      resolve(sha256);
-    });
+    return ampCurlUrlNodeSha256(str);
   }
-}
-
-/**
- * @param {string} buffer
- * @return {string}
- * @private
- */
-function hex_(buffer) {
-  let hexCodes = [];
-  const view = new DataView(buffer);
-  for (let i = 0; i < view.byteLength; i += 4) {
-    // Using getUint32 reduces the number of iterations needed (we process 4 bytes each time)
-    const value = view.getUint32(i);
-    // toString(16) will give the hex representation of the number without padding
-    const stringValue = value.toString(16);
-    // Use concatenation and slice for padding
-    const padding = '00000000';
-    const paddedValue = (padding + stringValue).slice(-padding.length);
-    hexCodes.push(paddedValue);
-  }
-
-  // Join all the hex strings into one
-  return hexCodes.join('');
 }
 
 /**
