+++ packages/cache-url/lib/AmpCurlUrlNodeSha256.js
@@ -1,0 +1,29 @@
+/**
+ * Copyright 2018 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * @param {string} str The string to convert to sha256
+ * @return {!Promise<string>}
+ * @private
+ */
+export default function ampCurlUrlNodeSha256(str) {
+  const buffer = Buffer.from(str, 'utf-8');
+  const crypto = require('crypto');
+  return new Promise((resolve) => {
+    const sha256 = crypto.createHash('sha256').update(buffer).digest('hex');
+    resolve(sha256);
+  });
+}
