+++ packages/cache-url/package.json
@@ -42,6 +42,8 @@
   },
   "devDependencies": {
     "@ampproject/rollup-plugin-closure-compiler": "0.8.2",
+    "@babel/core": "^7.1.2",
+    "babel-plugin-filter-imports": "^2.0.3",
     "eslint": "5.7.0",
     "eslint-config-google": "0.10.0",
     "jasmine": "3.2.0",
@@ -50,11 +52,10 @@
     "karma-jasmine": "1.1.2",
     "npm-run-all": "4.1.3",
     "rollup": "0.66.6",
+    "rollup-plugin-babel": "^4.0.3",
     "rollup-plugin-commonjs": "9.2.0",
     "rollup-plugin-filesize": "5.0.1",
-    "rollup-plugin-ignore": "1.0.4",
     "rollup-plugin-json": "3.1.0",
-    "rollup-plugin-node-builtins": "2.1.2",
     "rollup-plugin-node-resolve": "3.4.0",
     "rollup-plugin-serve": "0.6.0",
     "semver": "5.6.0"
